Fix: tests
